* commit 'fe6b2366600d53c6d7607d576a9ca192c6ba70b1':
  Make domain a uint8_t.

+++ capdl-loader-app/include/capdl.h
@@ -220,7 +220,7 @@ typedef struct {
     const char *elf_name;
     const seL4_Word *init;
     seL4_Word init_sz;
-    uint32_t domain;
+    uint8_t domain;
     seL4_CPtr fault_ep;
 } CDL_TCBExtraInfo;
